Fix celery structure (#3)
