+++ backend/Dockerfile.dev
@@ -10,4 +10,4 @@ RUN pip install --no-cache-dir -r requirements/dev.txt
 
 COPY . .
 
-CMD flask run --host=0.0.0.0 --port=5000 --reload & celery -A celery_app.celery worker --loglevel=info --autoscale=10,2
+CMD flask run --host=0.0.0.0 --port=5000 --reload & celery -A wsgi.celery worker --loglevel=info --autoscale=10,2
