+++ backend/wsgi.py
@@ -4,8 +4,10 @@ from flask_migrate import Migrate
 from app import create_app, db, mail
 from app.models import User, Role, Card, Transaction, FinancialReport, Permission
 
+
 app = create_app(os.getenv('FLASK_CONFIG') or 'default')
 migrate = Migrate(app, db)
+celery = app.celery
 
 
 @app.shell_context_processor
